flytt filter-logikk for avtaler til egen fil
