+++ frontend/mr-admin-flate/src/api/atoms.ts
@@ -1,20 +1,12 @@
-import { ARRANGORER_PAGE_SIZE, AVTALE_PAGE_SIZE } from "@/constants";
+import { ARRANGORER_PAGE_SIZE } from "@/constants";
 import { SortState } from "@navikt/ds-react";
-import { atom, WritableAtom } from "jotai";
+import { atom } from "jotai";
 import {
-  atomFamily,
   atomWithStorage,
   createJSONStorage,
   unstable_withStorageValidator as withStorageValidator,
 } from "jotai/utils";
-import {
-  Avtalestatus,
-  Avtaletype,
-  OppgaveType,
-  SorteringArrangorer,
-  SorteringAvtaler,
-  SorteringTiltakstyper,
-} from "@mr/api-client-v2";
+import { OppgaveType, SorteringArrangorer, SorteringTiltakstyper } from "@mr/api-client-v2";
 import { z, ZodType } from "zod";
 
 export function createSorteringProps(sortItems: z.ZodType) {
@@ -54,55 +46,6 @@ export const tiltakstypeFilterAtom = atomWithStorage<TiltakstypeFilterType>(
   "tiltakstype-filter",
   defaultTiltakstypeFilter,
   tiltakstypeFilterStorage,
-  { getOnInit: true },
-);
-
-export const AvtaleFilterSchema = z.object({
-  sok: z.string(),
-  statuser: z.custom<Avtalestatus>().array(),
-  avtaletyper: z.custom<Avtaletype>().array(),
-  navRegioner: z.string().array(),
-  tiltakstyper: z.string().array(),
-  sortering: createSorteringProps(z.custom<SorteringAvtaler>()),
-  arrangorer: z.string().array(),
-  visMineAvtaler: z.boolean(),
-  personvernBekreftet: z.boolean().optional(),
-  page: z.number(),
-  pageSize: z.number(),
-  lagretFilterIdValgt: z.string().optional(),
-});
-
-export type AvtaleFilterType = z.infer<typeof AvtaleFilterSchema>;
-
-export const defaultAvtaleFilter: AvtaleFilterType = {
-  sok: "",
-  statuser: [Avtalestatus.AKTIV],
-  avtaletyper: [],
-  navRegioner: [],
-  tiltakstyper: [],
-  sortering: {
-    sortString: SorteringAvtaler.NAVN_ASCENDING,
-    tableSort: {
-      orderBy: "navn",
-      direction: "ascending",
-    },
-  },
-  arrangorer: [],
-  visMineAvtaler: false,
-  personvernBekreftet: undefined,
-  page: 1,
-  pageSize: AVTALE_PAGE_SIZE,
-  lagretFilterIdValgt: undefined,
-};
-
-const avtaleFilterStorage = withStorageValidator(createFilterValidator(AvtaleFilterSchema))(
-  createJSONStorage(() => sessionStorage),
-);
-
-export const avtaleFilterAtom = atomWithStorage<AvtaleFilterType>(
-  "avtale-filter",
-  defaultAvtaleFilter,
-  avtaleFilterStorage,
   { getOnInit: true },
 );
 
@@ -163,31 +106,10 @@ export const oppgaverFilterAtom = atomWithStorage<OppgaverFilterType>(
   { getOnInit: true },
 );
 
-const avtalerForTiltakstypeFilterAtomFamily = atomFamily<
-  AvtaleFilterType,
-  WritableAtom<AvtaleFilterType, [newValue: AvtaleFilterType], void>
->(
-  (filter: AvtaleFilterType) => {
-    return atomWithStorage(
-      `avtale-filter-${filter.tiltakstyper[0]}`,
-      filter,
-      createJSONStorage(() => sessionStorage),
-    );
-  },
-  (a, b) => a.tiltakstyper[0] === b.tiltakstyper[0],
-);
-
-export function getAvtalerForTiltakstypeFilterAtom(tiltakstypeId: string) {
-  const defaultFilterValue = { ...defaultAvtaleFilter, tiltakstyper: [tiltakstypeId] };
-  const filterAtom = avtalerForTiltakstypeFilterAtomFamily(defaultFilterValue);
-  return { defaultFilterValue, filterAtom };
-}
-
 export const gjennomforingDetaljerTabAtom = atom<"detaljer" | "redaksjonelt-innhold">("detaljer");
 
 export const avtaleDetaljerTabAtom = atom<
   "detaljer" | "okonomi" | "personvern" | "redaksjonelt-innhold"
 >("detaljer");
 
-export const avtaleFilterAccordionAtom = atom<string[]>(["region"]);
 export const oppgaverFilterAccordionAtom = atom<string[]>(["type", "regioner"]);
