+++ frontend/mr-admin-flate/src/components/tabell/AvtaleTabell.tsx
@@ -1,13 +1,7 @@
-import { AvtaleFilterType } from "@/api/atoms";
 import { EksporterTabellKnapp } from "@/components/eksporterTabell/EksporterTabellKnapp";
 import { TabellWrapper } from "@/components/tabell/TabellWrapper";
-import {
-  capitalizeEveryWord,
-  createQueryParamsForExcelDownloadForAvtale,
-  formaterDato,
-  formaterNavEnheter,
-} from "@/utils/Utils";
-import { AvtalerService, SorteringAvtaler } from "@mr/api-client-v2";
+import { capitalizeEveryWord, formaterDato, formaterNavEnheter } from "@/utils/Utils";
+import { SorteringAvtaler } from "@mr/api-client-v2";
 import { Lenke } from "@mr/frontend-common/components/lenke/Lenke";
 import { ToolbarContainer } from "@mr/frontend-common/components/toolbar/toolbarContainer/ToolbarContainer";
 import { ToolbarMeny } from "@mr/frontend-common/components/toolbar/toolbarMeny/ToolbarMeny";
@@ -18,6 +12,8 @@ import { Laster } from "../laster/Laster";
 import { PagineringContainer } from "../paginering/PagineringContainer";
 import { PagineringsOversikt } from "../paginering/PagineringOversikt";
 import { AvtaleStatusTag } from "../statuselementer/AvtaleStatusTag";
+import { AvtaleFilterType } from "@/pages/avtaler/filter";
+import { downloadAvtalerAsExcel } from "@/api/avtaler/downloadAvtalerAsExcel";
 
 interface Props {
   filter: AvtaleFilterType;
@@ -34,19 +30,13 @@ export function AvtaleTabell({ filter, updateFilter, tagsHeight, filterOpen }: P
 
   const link = createRef<HTMLAnchorElement>();
 
-  async function lastNedFil(filter: AvtaleFilterType) {
-    const query = createQueryParamsForExcelDownloadForAvtale(filter);
-    const { data } = await AvtalerService.lastNedAvtalerSomExcel(query);
-    return data;
-  }
-
   async function lastNedExcel() {
     setLasterExcel(true);
     if (excelUrl) {
       setExcelUrl("");
     }
 
-    const excelFil = await lastNedFil(filter);
+    const excelFil = await downloadAvtalerAsExcel(filter);
     const url = URL.createObjectURL(excelFil);
     setExcelUrl(url);
     setLasterExcel(false);
